+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## 1.0.0
+## 0.1.0
 
-- Initial version.
+- Wikipedia Search, Summary with lang parameter
+
+## 0.2.0
+
+- Wikidata GetEntities
+
+## 0.3.0
+
+- Wikipedia Get Page Props
+